IT IS WORK

+++ ChatBot.Backend/app/models/project.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, TIMESTAMP, Text
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey, TIMESTAMP, Text
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.db.database import Base
@@ -8,7 +8,7 @@ class Project(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
-    user_id = Column(Integer, ForeignKey("users.telegram_id"))
+    user_id = Column(BigInteger, ForeignKey("users.telegram_id"))
     created_at = Column(TIMESTAMP, default=datetime.utcnow)
     system_prompt = Column(Text, nullable=True)
